requirements.txt
alembic
fastapi
uvicorn
sqlalchemy
databases
pydantic
PyJWT
bcrypt
passlib
asyncpg
pandas
python-dotenv
fastapi-mail
redis
python-multipart 
pillow
dropbox